+++ AudioKit/iOS/AudioKit/AudioKit.playground/contents.xcplayground
@@ -25,7 +25,6 @@
         <page name='Delay'/>
         <page name='Distortion'/>
         <page name='Drawbar Organ'/>
-        <page name='Drawbar Organ 2'/>
         <page name='Dripping Sounds'/>
         <page name='Drum Synthesizers'/>
         <page name='Dry Wet Mixer'/>
